changed data type of distance amount

+++ FitbitLogin/db/migrate/20150407211316_create_distances.rb
@@ -1,7 +1,7 @@
 class CreateDistances < ActiveRecord::Migration
   def change
     create_table :distances do |t|
-      t.integer :distanceAmount
+      t.float :distanceAmount
       t.date :distanceDate
 	  t.integer :user_id
